Cleanup examples
