+++ examples/colors.py
